Modify to do list interactive
